requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
google-auth==2.37.0
idna==3.10
msgpack==1.1.0
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
redis==5.2.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
urllib3==2.3.0
